Fix InterpolationSearch subtraction overflow with negative array values
- Cast key and arr[high] to long before subtraction to prevent int
  overflow when array contains values spanning a wide range (e.g.
  -2B to 2B)

+++ src/main/java/com/practice/dsalgo/searching/InterpolationSearch.java
@@ -129,7 +129,7 @@ public class InterpolationSearch {
 			}
 
 			// Estimate the position using interpolation formula
-			int pos = low + (int)(((long)(key - arr[low]) * (high - low)) / (arr[high] - arr[low]));
+			int pos = low + (int)((((long)key - arr[low]) * (high - low)) / ((long)arr[high] - arr[low]));
 
 			if (arr[pos] == key) {
 				return pos;
